+++ R/plot_correlation.r
@@ -8,6 +8,7 @@
 #' @param ggtheme complete ggplot2 themes. The default is \link{theme_gray}.
 #' @param theme_config a list of configurations to be passed to \link{theme}.
 #' @param \dots other arguments to be passed to \link{cor}.
+#' @return invisibly return the ggplot object
 #' @keywords plot_correlation
 #' @details For discrete features, the function first dummifies all categories, then calculates the correlation matrix (see \link{cor}) and plots it.
 #' @details To change default font family and size, you may pass \code{base_size} and \code{base_family} to \code{ggtheme} options, e.g., \code{ggtheme = theme_gray(base_size = 15, base_family = "serif")}
@@ -22,7 +23,6 @@
 #'
 #' # Plot correlation heatmap
 #' plot_correlation(diamonds)
-#' plot_correlation(diamonds, maxcat = 5)
 #' plot_correlation(diamonds, type = "c")
 #' plot_correlation(diamonds, type = "d")
 
@@ -56,16 +56,19 @@ plot_correlation <- function(data, type = c("all", "discrete", "continuous"), ma
   ## Calculate correlation and melt into tidy data format
   plot_data <- reshape2::melt(cor(final_data, ...))
   ## Create ggplot object
-  plot <- ggplot(plot_data, aes(x = Var1, y = Var2, fill = value)) +
+  output <- ggplot(plot_data, aes(x = Var1, y = Var2, fill = value)) +
     geom_tile() +
     scale_fill_gradient2("Correlation Meter", limits = c(-1, 1), low = "#0571b0", high = "#ca0020", space = "Lab") +
-    xlab("Features") + ylab("Features") +
-    ggtitle(label = title) +
-    ggtheme +
-    do.call(theme, theme_config)
+    xlab("Features") + ylab("Features")
   if (ncol(final_data) <= 20) {
-    plot <- plot + geom_text(aes(label = round(value, 2)))
+  	output <- output + geom_text(aes(label = round(value, 2)))
   }
-  ## Print plot object
-  print(plot)
+  ## Plot object
+  class(output) <- c("single", class(output))
+  plotDataExplorer(
+  	obj = output,
+  	title = title,
+  	ggtheme = ggtheme,
+  	theme_config = theme_config
+  )
 }