+++ R/plot_histogram.r
@@ -14,10 +14,8 @@
 #' @details \code{theme_config} argument expects all inputs to be wrapped in a list object, e.g., to change the text color: \code{theme_config = list("text" = element_text(color = "blue"))}
 #' @import data.table
 #' @import ggplot2
-#' @import gridExtra
-#' @importFrom stats na.omit
+#' @importFrom stats setNames
 #' @importFrom parallel mclapply
-#' @importFrom parallel detectCores
 #' @export plot_histogram
 #' @seealso \link{geom_histogram} \link{plot_density}
 #' @examples
@@ -53,20 +51,20 @@ plot_histogram <- function(data, title = NULL, ggtheme = theme_gray(), theme_con
 	plot_list <- mclapply(
 		setNames(seq_along(continuous), names(continuous)),
 		function(j) {
-			x <- na.omit(continuous[, j, with = FALSE])
+			x <- continuous[, j, with = FALSE]
 			ggplot(x, aes_string(x = names(x))) +
-				geom_histogram(bins = 30L, ...) +
+				geom_histogram(na.rm = TRUE, ...) +
 				ylab("Frequency") +
 				ggtheme +
 				do.call(theme, theme_config)
 		},
 		mc.preschedule = TRUE,
 		mc.silent = TRUE,
-		mc.cores = detectCores() - 1L
+		mc.cores = .getCores()
 	)
 	## Plot objects
-	class(plot_list) <- c(class(plot_list), "DataExplorerGrid")
-	plot.DataExplorerGrid(
+	class(plot_list) <- c("grid", class(plot_list))
+	plotDataExplorer(
 		obj_list = plot_list,
 		page_layout = layout,
 		nrow = nrow,