+++ NEWS.md
@@ -1,14 +1,25 @@
 # Changelog
 
-### DataExplorer 0.7.0
+### DataExplorer 0.6.1.9000
 #### Enhancements
+* [#42](https://github.com/boxuancui/DataExplorer/issues/42): Applied S3 methods for plotting functions.
+* [#78](https://github.com/boxuancui/DataExplorer/issues/78): All ggplot objects from `plot_*` are now invisibly returned. As a result, extracted `profile_missing` from `plot_missing` for missing value profiles.
 * [#83](https://github.com/boxuancui/DataExplorer/issues/83): Removed all deprecated functions.
+* [#85](https://github.com/boxuancui/DataExplorer/issues/85): Users can now specify number of rows/columns for plot page layout.
+* Plot objects are now constructed with `parallel::mclapply`.
+
+#### Bug Fixes
+* [#82](https://github.com/boxuancui/DataExplorer/issues/82): Removed hard-coded scales from all plot functions.
+* Fixed a bug causing wrong column indices in `split_columns`.
+* Fixed a bug using standard deviation instead of variance in `plot_prcomp`.
+
+### DataExplorer 0.6.1
+#### Enhancements
 * Updated vignette for better clarity.
 * Added better error handler for `plot_prcomp`.
 
 #### Bug Fixes
 * [#69](https://github.com/boxuancui/DataExplorer/issues/69): Fixed bug causing `create_report` failure (specifically from `plot_prcomp`) when `y` is specified.
-* [#82](https://github.com/boxuancui/DataExplorer/issues/82): Removed hard-coded scales from `plot_histogram`.
 * Added more unit tests for `create_report` and `plot_prcomp`.
 
 ### DataExplorer 0.6.0